Fixed various bugs
Seems to be working consistently and correctly for uint96 and uint128

Prior to these fixes, the code failed when the modulus was significantly
smaller than the specified bitlength. In particular, when the modulus is
small the hi-order word of mu might be greater than one.

Hopefully this code works properly for larger word sizes, too!